test: move engine-dependent client integration tests to the root module (keep the client module engine-free)
