+++ client_vector_lifecycle_test.go
@@ -1,18 +1,19 @@
 // SPDX-License-Identifier: Apache-2.0
-package client
+package rostam
 
 import (
 	"context"
 	"errors"
 	"testing"
 
+	"github.com/rostamlabs/rostam/client"
 	"github.com/rostamlabs/rostam/sdk/vtypes"
 )
 
 func TestCollectionCreateAndDrop(t *testing.T) {
 	addr, stop := startTestStack(t)
 	defer stop()
-	c, err := New(Config{Servers: []string{addr}})
+	c, err := client.New(client.Config{Servers: []string{addr}})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -20,7 +21,7 @@ func TestCollectionCreateAndDrop(t *testing.T) {
 
 	ctx := context.Background()
 	col := c.Collection("posts")
-	if err := col.Create(ctx, CreateRequest{
+	if err := col.Create(ctx, client.CreateRequest{
 		Dim:      4,
 		Metric:   vtypes.Cosine,
 		FullText: &vtypes.FullTextConfig{Analyzer: "english"},
@@ -28,8 +29,8 @@ func TestCollectionCreateAndDrop(t *testing.T) {
 		t.Fatalf("Create: %v", err)
 	}
 	// Second create of the same name must surface ErrCollectionExists.
-	err = col.Create(ctx, CreateRequest{Dim: 4, Metric: vtypes.Cosine})
-	if !errors.Is(err, ErrCollectionExists) {
+	err = col.Create(ctx, client.CreateRequest{Dim: 4, Metric: vtypes.Cosine})
+	if !errors.Is(err, client.ErrCollectionExists) {
 		t.Fatalf("second Create: want ErrCollectionExists, got %v", err)
 	}
 	if err := col.Drop(ctx); err != nil {
