+++ client_vector_write_test.go
@@ -1,23 +1,24 @@
 // SPDX-License-Identifier: Apache-2.0
-package client
+package rostam
 
 import (
 	"context"
 	"testing"
 
+	"github.com/rostamlabs/rostam/client"
 	"github.com/rostamlabs/rostam/sdk/vtypes"
 )
 
-func mustCollection(t *testing.T) (*Collection, func()) {
+func mustCollection(t *testing.T) (*client.Collection, func()) {
 	t.Helper()
 	addr, stop := startTestStack(t)
-	c, err := New(Config{Servers: []string{addr}})
+	c, err := client.New(client.Config{Servers: []string{addr}})
 	if err != nil {
 		stop()
 		t.Fatal(err)
 	}
 	col := c.Collection("posts")
-	if err := col.Create(context.Background(), CreateRequest{Dim: 4, Metric: vtypes.Cosine}); err != nil {
+	if err := col.Create(context.Background(), client.CreateRequest{Dim: 4, Metric: vtypes.Cosine}); err != nil {
 		_ = c.Close()
 		stop()
 		t.Fatal(err)
@@ -30,7 +31,7 @@ func TestUpsertThenDelete(t *testing.T) {
 	defer cleanup()
 	ctx := context.Background()
 
-	err := col.Upsert(ctx, WriteRequest{
+	err := col.Upsert(ctx, client.WriteRequest{
 		ID:       1,
 		Vector:   []float32{0.1, 0.2, 0.3, 0.4},
 		Content:  "hello world",
@@ -39,7 +40,7 @@ func TestUpsertThenDelete(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Upsert: %v", err)
 	}
-	if err := col.Delete(ctx, DeleteRequest{ID: 1}); err != nil {
+	if err := col.Delete(ctx, client.DeleteRequest{ID: 1}); err != nil {
 		t.Fatalf("Delete: %v", err)
 	}
 }
@@ -49,15 +50,15 @@ func TestUpsertVersionConflict(t *testing.T) {
 	defer cleanup()
 	ctx := context.Background()
 
-	if err := col.Upsert(ctx, WriteRequest{ID: 7, Vector: []float32{1, 0, 0, 0}}); err != nil {
+	if err := col.Upsert(ctx, client.WriteRequest{ID: 7, Vector: []float32{1, 0, 0, 0}}); err != nil {
 		t.Fatalf("seed Upsert: %v", err)
 	}
 	// Expect a version that cannot match a fresh point.
-	err := col.Upsert(ctx, WriteRequest{
+	err := col.Upsert(ctx, client.WriteRequest{
 		ID: 7, Vector: []float32{0, 1, 0, 0},
 		ExpectedVersion: 999, HasExpectedVersion: true,
 	})
-	if err != ErrVersionConflict {
+	if err != client.ErrVersionConflict {
 		t.Fatalf("Upsert with stale version = %v, want ErrVersionConflict", err)
 	}
 }
@@ -67,7 +68,7 @@ func TestUpsertBatch(t *testing.T) {
 	defer cleanup()
 	ctx := context.Background()
 
-	pts := []PointInput{
+	pts := []client.PointInput{
 		{ID: 1, Vector: []float32{1, 0, 0, 0}, Content: "a"},
 		{ID: 2, Vector: []float32{0, 1, 0, 0}, Content: "b"},
 		{ID: 3, Vector: []float32{0, 0, 1, 0}, Content: "c"},
@@ -84,11 +85,11 @@ func TestInsertRejectsContent(t *testing.T) {
 
 	// The insert wire op carries no content field, so Insert must refuse it
 	// rather than silently drop it.
-	if err := col.Insert(ctx, WriteRequest{ID: 1, Vector: []float32{1, 0, 0, 0}, Content: "x"}); err == nil {
+	if err := col.Insert(ctx, client.WriteRequest{ID: 1, Vector: []float32{1, 0, 0, 0}, Content: "x"}); err == nil {
 		t.Fatal("Insert with Content = nil error, want rejection")
 	}
 	// A plain Insert (no content) still works.
-	if err := col.Insert(ctx, WriteRequest{ID: 2, Vector: []float32{0, 1, 0, 0}}); err != nil {
+	if err := col.Insert(ctx, client.WriteRequest{ID: 2, Vector: []float32{0, 1, 0, 0}}); err != nil {
 		t.Fatalf("Insert without content: %v", err)
 	}
 }
